+++ conreality_headset/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 0.1.2 - 2019-07-11
+### Changed
+- Fixed a bug in `Headset.isConnected`
 ### Added
 - `Headset.canSpeak` static property
 - `Headset.speak()` static method